server/requirements.txt
annotated-types==0.7.0

anyio==4.6.2.post1
asgiref==3.8.1
beautifulsoup4==4.12.3
boto3==1.37.37
botocore==1.37.37
cachetools==5.5.2
certifi==2024.8.30
charset-normalizer==3.4.2
click==8.1.7
distro==1.9.0
Django==5.1.3
django-appconf==1.0.6
django-compressor==4.5.1
django-compressor-autoprefixer==0.1.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
jiter==0.7.0
jmespath==1.0.1
lxml==5.3.0
Markdown==3.7
oauthlib==3.2.2
openai==1.54.3
packaging==24.2
proto-plus==1.26.1
protobuf==6.30.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
rcssmin==1.1.2
requests==2.32.3
requests-oauthlib==2.0.0
rjsmin==1.2.2
rsa==4.9.1
s3transfer==0.11.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
tqdm==4.67.0
typing-inspection==0.4.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.32.0
uvicorn-worker==0.2.0
whitenoise==6.8.2
yt-dlp==2025.4.30